+++ Sources/BraintreeCore/Public/BraintreeCore/BraintreeCore.h
@@ -8,8 +8,6 @@ FOUNDATION_EXPORT const unsigned char BraintreeCoreVersionString[];
 
 #if __has_include(<Braintree/BraintreeCore.h>)
 #import <Braintree/BTAPIClient.h>
-#import <Braintree/BTBinData.h>
-#import <Braintree/BTClientToken.h>
 #import <Braintree/BTEnums.h>
 #import <Braintree/BTViewControllerPresentingDelegate.h>
 #import <Braintree/BTPreferredPaymentMethods.h>
@@ -17,8 +15,6 @@ FOUNDATION_EXPORT const unsigned char BraintreeCoreVersionString[];
 #import <Braintree/Braintree.h>
 #else
 #import <BraintreeCore/BTAPIClient.h>
-#import <BraintreeCore/BTBinData.h>
-#import <BraintreeCore/BTClientToken.h>
 #import <BraintreeCore/BTEnums.h>
 #import <BraintreeCore/BTViewControllerPresentingDelegate.h>
 #import <BraintreeCore/BTPreferredPaymentMethods.h>